Use int instead of long for xp reward config
Changed to long because `client.getOverallXp()` returns long, but

1) Apparently, this change wasn't necessary
2) It seems config items cannot be `long` (the input field does not show up). That being said, a long is really not necessary for this field

+++ src/main/java/com/cworldender/PersonalCurrencyTrackerConfig.java
@@ -268,7 +268,7 @@ public interface PersonalCurrencyTrackerConfig extends Config
 		description = "Reward Coins every x XP. If set to 0, no rewards will take place (But the accumulated XP will still be counted).",
 		section = xpSection
 	)
-	default long xpRewardInterval(){ return 0; }
+	default int xpRewardInterval(){ return 0; }
 
 	@ConfigItem(
 		keyName = "xpReward",
